added a utils file for builder
